+++ defaults.php
@@ -332,23 +332,45 @@ c::set(array(
    */
   'fileinfo' => array(
 
-    // content files
-    'txt'      => array('type' => 'content', 'mime' => 'text/plain'),
-    'mdown'    => array('type' => 'content', 'mime' => 'text/plain'),
-    'md'       => array('type' => 'content', 'mime' => 'text/plain'),
-    'markdown' => array('type' => 'content', 'mime' => 'text/plain'),
-
     // images
     'jpg'      => array('type' => 'image', 'mime' => 'image/jpeg'),
     'jpeg'     => array('type' => 'image', 'mime' => 'image/jpeg'),
     'gif'      => array('type' => 'image', 'mime' => 'image/gif'),
     'png'      => array('type' => 'image', 'mime' => 'image/png'),
+    'svg'      => array('type' => 'image', 'mime' => 'image/svg+xml'),
+    'ico'      => array('type' => 'image', 'mime' => 'image/x-icon'),
+    'tif'      => array('type' => 'image', 'mime' => 'image/tiff'),
+    'tiff'     => array('type' => 'image', 'mime' => 'image/tiff'),
+    'bmp'      => array('type' => 'image', 'mime' => 'image/bmp'),
 
     // documents
+    'txt'      => array('type' => 'document', 'mime' => 'text/plain'),
+    'mdown'    => array('type' => 'document', 'mime' => 'text/plain'),
+    'md'       => array('type' => 'document', 'mime' => 'text/plain'),
+    'markdown' => array('type' => 'document', 'mime' => 'text/plain'),
     'pdf'      => array('type' => 'document', 'mime' => 'application/pdf'),
     'doc'      => array('type' => 'document', 'mime' => 'application/msword'),
+    'docx'     => array('type' => 'document', 'mime' => 'application/msword'),
     'xls'      => array('type' => 'document', 'mime' => 'application/msexcel'),
+    'xlsx'     => array('type' => 'document', 'mime' => 'application/msexcel'),
     'ppt'      => array('type' => 'document', 'mime' => 'application/mspowerpoint'),
+
+    // archives
+    'zip'      => array('type' => 'archive', 'mime' => 'application/zip'),
+    'tar'      => array('type' => 'archive', 'mime' => 'application/x-tar'),
+    'gz'       => array('type' => 'archive', 'mime' => 'application/x-gzip'),
+    'gzip'     => array('type' => 'archive', 'mime' => 'application/x-gzip'),
+    'tgz'      => array('type' => 'archive', 'mime' => 'application/gnutar'),
+
+    // code
+    'js'       => array('type' => 'code', 'mime' => 'application/javascript'),
+    'css'      => array('type' => 'code', 'mime' => 'text/css'),
+    'scss'     => array('type' => 'code', 'mime' => 'text/css'),
+    'htm'      => array('type' => 'code', 'mime' => 'text/html'),
+    'html'     => array('type' => 'code', 'mime' => 'text/html'),
+    'php'      => array('type' => 'code', 'mime' => 'text/php'),
+    'xml'      => array('type' => 'code', 'mime' => 'application/xml'),
+    'json'     => array('type' => 'code', 'mime' => 'application/json'),
 
     // videos
     'mov'      => array('type' => 'video', 'mime' => 'video/quicktime'),
@@ -359,9 +381,13 @@ c::set(array(
     'flv'      => array('type' => 'video', 'mime' => 'video/x-flv'),
     'swf'      => array('type' => 'video', 'mime' => 'application/x-shockwave-flash'),
     'mp4'      => array('type' => 'video', 'mime' => 'video/mp4'),
+    'mv4'      => array('type' => 'video', 'mime' => 'video/mv4'),
 
-    // sounds
-    'mp3'      => array('type' => 'sound', 'mime' => 'audio/mp3'),
+    // audio
+    'mp3'      => array('type' => 'audio', 'mime' => 'audio/mp3'),
+    'wav'      => array('type' => 'audio', 'mime' => 'audio/wav'),
+    'aif'      => array('type' => 'audio', 'mime' => 'audio/aiff'),
+    'aiff'     => array('type' => 'audio', 'mime' => 'audio/aiff'),
 
   ),